requirements.txt
Bottleneck==1.3.4
certifi==2021.5.30
cycler==0.11.0
et-xmlfile==1.1.0
fonttools==4.25.0
kiwisolver==1.3.1
matplotlib==3.5.1
mkl-fft==1.3.1
mkl-random==1.2.2
mkl-service==2.4.0
munkres==1.1.4
numexpr==2.8.1
numpy==1.21.5
openpyxl==3.0.9
packaging==21.3
pandas==1.4.2
Pillow==9.0.1
pip==21.2.4
pyparsing==3.0.4
python-dateutil==2.8.2
pytz==2021.3
PyYAML==6.0
setuptools==61.2.0
six==1.16.0
tornado==6.1
wheel==0.37.1
XlsxWriter==3.0.3
